+++ astrodyl/resources/tmp/description.txt
@@ -1,0 +1,23 @@
+MJD: 1-9
+Tel: 11-17
+ruJy: 19-24
+e_ruJy: 26-30
+rmag: 32-36
+e_rmag: 38-41
+iuJy: 43-48
+e_iuJy: 50-54
+l_imag: 56-56
+imag: 57-61
+e_imag: 63-66
+zuJy: 68-73
+e_zuJy: 75-80
+l_zmag: 82-82
+zmag: 83-87
+e_zmag: 89-92
+yuJy: 94-99
+e_yuJy: 101-105
+l_ymag: 107-107
+ymag: 108-112
+e_ymag: 114-117
+ExpTime: 119-150
+Method: 152-161
