+++ astrodyl/resources/tmp/sources.txt
@@ -1,0 +1,15 @@
+https://www.swift.ac.uk/xrt_products/index.php?year=2007
+https://www.swift.ac.uk/xrt_curves/docs.php#products
+
+A comparative study of outflow structures of two classes of gamma-ray bursts
+A DECADE OF SHORT-DURATION GAMMA-RAY BURST BROADBAND AFTERGLOWS: ENERGETICS, CIRCUMBURST DENSITIES, AND JET OPENING ANGLES
+GAMMA-RAYBURSTS ARE OBSERVED OFF-AXIS
+https://cosmo.nyu.edu/afterglowlibrary/scalefitdata.html
+
+
+
+
+https://www.zillow.com/homedetails/530-Lambert-Ln-Fuquay-Varina-NC-27526/339885896_zpid/
+
+
+GRB080319B
