+++ astrodyl/skynet/defns/telescope.py
