--- grb/resources/cuts/break_times.json
@@ -1,17 +1,0 @@
-{
-	"GRB050525A": [6012],
-	"GRB050922C": [2767],
-	"GRB080319B": [81, 476, 498, 2259],
-	"GRB080413B": [2858],
-	"GRB090424":  [119, 248, 1758, 1837, 282000],
-	"GRB090618":  [221, 390, 1607, 9506, 360000],
-	"GRB111228A": [158, 408, 16900, 29000, 52800],
-	"GRB130612A": [],
-	"GRB131030A": [96, 330, 6427],
-	"GRB140506A": [897, 13000],
-	"GRB160131A": [114, 169, 86000],
-	"GRB171010A": [70000, 170000],
-	"GRB210905A": [98, 279, 60000],
-	"GRB220101A": [119, 79000],
-	"GRB221009A": [32700, 38700, 45700, 135000]
-}
