--- grb/resources/grbs/GRB080319B/bloom_gemini.txt
@@ -1,27 +1,0 @@
-Table 7 
-Gemini Observations of GRB 080319B
-
-
-t_mid^a (s)	Filter	Exp. Time (s)	Mag.^b	Flux^b (muJy)	
-89309	g	5 x 180	20.670 +or- 0.100	19.59 +or- 1.72	
-90356	r	5 x 180	20.520 +or- 0.060	22.49 +or- 1.21	
-91307	i	5 x 180	20.380 +or- 0.050	25.59 +or- 1.15	
-92470	z	5 x 180	20.310 +or- 0.050	27.93 +or- 1.26	
-174960	r	5 x 180	21.510 +or- 0.060	9.037 +or- 0.486	
-520390	g	6 x 180	23.450 +or- 0.190	1.514 +or- 0.243	
-521589	r	6 x 180	23.340 +or- 0.090	1.675 +or- 0.133	
-522792	i	6 x 180	22.900 +or- 0.060	2.512 +or- 0.135	
-523990	z	6 x 180	22.940 +or- 0.090	2.477 +or- 0.197	
-780120	r	6 x 180	23.670 +or- 0.070	1.236 +or- 0.077	
-781560	i	6 x 180	23.280 +or- 0.060	1.770 +or- 0.095	
-1209960	g	6 x 180	24.440 +or- 0.080	0.6081 +or- 0.0432	
-1211220	r	6 x 180	24.100 +or- 0.080	0.8318 +or- 0.0591	
-1212480	i	6 x 180	23.810 +or- 0.070	1.086 +or- 0.068	
-2286720	g	6 x 180	25.860 +or- 0.110^x	0.1644 +or- 0.0158^x	
-2287800	r	6 x 180	25.050 +or- 0.090^x	0.3467 +or- 0.0276^x	
-2285280	i	6 x 180	24.360 +or- 0.080^x	0.6546 +or- 0.0465^x	
-2531160	z	3 x 180	23.940 +or- 0.130^x	0.9863 +or- 0.1113^x	
-Notes.
-a Exposure midtime, measured from the Swift trigger (UTC 06:12:49).
-b Observed value; not corrected for Galactic extinction.
-x Point not used in modeling.
