--- grb/resources/grbs/GRB080319B/bloom_kait.txt
@@ -1,154 +1,0 @@
-Title: Observations of the Naked-Eye GRB 080319B: Implications of Nature's 
-       Brightest Explosion  
-Authors: Bloom J.S., Perley D.A., Li W., Butler N.R., Miller A.A., Kocevski D.,
-         Kann D.A., Foley R.J., Chen H.-W., Filippenko A.V., Starr D.L., 
-         Macomber .B., Prochaska J.X., Chornock R., Poznanski D., Klose S., 
-         Skrutskie M.F., Lopez S., Hall P., Glazebrook K., Blake C.H. 
-Table: KAIT Observations of GRB 080319B
-================================================================================
-Byte-by-byte Description of file: apj292633t4_mrt.txt
---------------------------------------------------------------------------------
-   Bytes Format Units   Label  Explanations
---------------------------------------------------------------------------------
-   1-  7 F7.1   s       Tmid   Exposure mid-time (1)
-   9- 13 A5     ---     Filt   Filter used in observation
-  15- 19 F5.1   s       Exp    Exposure time  
-  21- 26 F6.3   mag     mag    Observed magnitude in Filt (2)
-  28- 32 F5.3   mag   e_mag    Uncertainty in mag
-  34- 39 F6.1   uJy     Flux   Observed flux density in Filt (2)
-  41- 45 F5.1   uJy   e_Flux   Uncertainty in Flux 
---------------------------------------------------------------------------------
-Note (1): Measured from the Swift trigger (UTC 06:12:49).
-Note (2): Not corrected for Galactic extinction.
---------------------------------------------------------------------------------
- 1217.0 clear  20.0 14.116 0.008 7008.1  51.4 
- 1308.0 clear  20.0 14.261 0.007 6132.0  39.4 
- 1404.0 clear  20.0 14.424 0.011 5277.2  53.2 
- 1495.0 clear  20.0 14.548 0.013 4707.6  56.0 
- 1586.0 clear  20.0 14.651 0.011 4281.5  43.2 
- 1651.0 clear  20.0 14.735 0.010 3962.8  36.3 
- 1714.0 clear  20.0 14.798 0.010 3739.4  34.3 
- 1773.0 clear  20.0 14.901 0.013 3401.0  40.5 
- 1832.0 clear  20.0 14.947 0.011 3259.9  32.9 
- 1942.0 clear  20.0 15.051 0.010 2962.1  27.2 
- 2001.0 clear  20.0 15.069 0.011 2913.4  29.4 
- 2128.0 clear  20.0 15.209 0.010 2560.9  23.5 
- 2188.0 clear  20.0 15.259 0.011 2445.7  24.7 
- 2308.0 clear  20.0 15.335 0.010 2280.3  20.9 
- 2432.0 clear  20.0 15.400 0.012 2147.8  23.6 
- 2557.0 clear  20.0 15.503 0.009 1953.4  16.1 
- 2678.0 clear  20.0 15.601 0.013 1784.8  21.2 
- 2798.0 clear  20.0 15.640 0.013 1721.9  20.5 
- 2922.0 clear  20.0 15.716 0.014 1605.5  20.6 
- 3047.0 clear  20.0 15.750 0.016 1556.0  22.8 
- 3168.0 clear  20.0 15.806 0.014 1477.7  18.9 
- 3289.0 clear  20.0 15.870 0.017 1393.2  21.6 
- 3414.0 clear  20.0 15.900 0.014 1355.2  17.4 
- 3534.0 clear  20.0 15.978 0.021 1261.2  24.2 
- 3655.0 clear  20.0 15.999 0.014 1237.1  15.8 
- 3780.0 clear  20.0 16.060 0.015 1169.5  16.0 
- 3901.0 clear  20.0 16.080 0.020 1148.2  21.0 
- 4026.0 clear  20.0 16.127 0.018 1099.5  18.1 
- 4146.0 clear  20.0 16.198 0.016 1029.9  15.1 
- 4267.0 clear  20.0 16.242 0.013  989.0  11.8 
- 4331.0 clear  20.0 16.262 0.017  971.0  15.1 
- 4390.0 clear  20.0 16.254 0.014  978.1  12.5 
- 4449.0 clear  20.0 16.278 0.017  956.8  14.9 
- 4508.0 clear  20.0 16.285 0.016  950.6  13.9 
- 4567.0 clear  20.0 16.334 0.014  908.7  11.6 
- 5015.0 clear  20.0 16.422 0.019  837.9  14.5 
- 5114.0 clear  20.0 16.471 0.011  800.9   8.1 
- 5316.0 clear  20.0 16.508 0.016  774.1  11.3 
- 5420.0 clear  20.0 16.541 0.017  750.9  11.7 
- 5619.0 clear  20.0 16.592 0.021  716.5  13.7 
- 5823.0 clear  20.0 16.633 0.012  689.9   7.6 
- 6023.0 clear  20.0 16.667 0.020  668.7  12.2 
- 6227.0 clear  20.0 16.739 0.017  625.7   9.7 
- 6426.0 clear  20.0 16.783 0.021  600.9  11.5 
- 6631.0 clear  20.0 16.773 0.025  606.5  13.8 
- 6830.0 clear  20.0 16.850 0.017  564.9   8.8 
- 7034.0 clear  20.0 16.892 0.022  543.5  10.9 
- 8820.0 clear 300.0 17.153 0.012  427.4   4.7 
-10028.0 clear 300.0 17.335 0.015  361.4   5.0 
-11271.0 clear 300.0 17.528 0.014  302.6   3.9 
-12477.0 clear 300.0 17.625 0.014  276.7   3.5 
-15537.0 clear 300.0 17.919 0.022  211.1   4.2 
-16744.0 clear 300.0 18.021 0.022  192.1   3.9 
- 2098.0 B      20.0 15.623 0.045 2327.0  94.5 
- 2249.0 B      20.0 15.750 0.041 2070.1  76.7 
- 2373.0 B      20.0 15.814 0.042 1951.6  74.1 
- 2546.5 B      40.0 15.911 0.037 1784.8  59.8 
- 2791.0 B      40.0 16.094 0.035 1508.0  47.8 
- 3097.3 B      60.0 16.305 0.031 1241.6  34.9 
- 3465.0 B      60.0 16.378 0.040 1160.9  42.0 
- 3771.5 B      40.0 16.574 0.043  969.2  37.6 
- 4075.7 B      60.0 16.602 0.045  944.5  38.3 
- 5493.7 B     120.0 17.036 0.035  633.3  20.1 
- 6117.3 B     120.0 17.173 0.036  558.2  18.2 
- 6720.7 B     120.0 17.412 0.068  447.9  27.2 
- 7789.0 B     300.0 17.518 0.023  406.3   8.5 
- 8984.0 B     300.0 17.737 0.030  332.0   9.0 
-10236.0 B     300.0 17.833 0.028  303.9   7.7 
-11435.0 B     300.0 18.065 0.044  245.5   9.7 
-14508.0 B     300.0 18.440 0.052  173.8   8.1 
-15707.0 B     300.0 18.490 0.058  166.0   8.6 
- 1157.0 V      20.0 14.345 0.022 6745.3 135.3 
- 1248.0 V      20.0 14.442 0.022 6168.8 123.7 
- 1339.0 V      20.0 14.651 0.022 5088.6 102.1 
- 1435.0 V      20.0 14.815 0.025 4375.2  99.6 
- 1526.0 V      20.0 14.939 0.026 3903.0  92.4 
- 2067.0 V      20.0 15.492 0.043 2345.3  91.1 
- 2218.0 V      20.0 15.631 0.047 2063.5  87.4 
- 2453.3 V      60.0 15.800 0.029 1766.0  46.5 
- 2820.0 V      60.0 15.921 0.036 1579.8  51.5 
- 3188.7 V      60.0 16.130 0.037 1303.2  43.7 
- 3556.3 V      60.0 16.336 0.043 1078.0  41.9 
- 3984.3 V      80.0 16.421 0.037  996.8  33.4 
- 5332.5 V      80.0 16.827 0.042  685.8  26.0 
- 5782.0 V      80.0 16.954 0.055  610.1  30.1 
- 6165.5 V      80.0 16.989 0.047  590.7  25.0 
- 6569.0 V      80.0 17.111 0.060  528.0  28.4 
- 6973.0 V      80.0 17.144 0.068  512.2  31.1 
- 8134.0 V     300.0 17.453 0.057  385.3  19.7 
- 9335.0 V     300.0 17.629 0.069  327.6  20.2 
-10580.0 V     300.0 17.776 0.019  286.2   5.0 
-11785.0 V     300.0 17.958 0.031  242.0   6.8 
-14852.0 V     300.0 18.223 0.029  189.6   5.0 
-16051.0 V     300.0 18.370 0.038  165.6   5.7 
- 1188.0 I      20.0 13.777 0.013 7502.4  89.3 
- 1279.0 I      20.0 13.879 0.012 6829.7  75.1 
- 1370.0 I      20.0 14.016 0.015 6020.0  82.6 
- 1465.0 I      20.0 14.176 0.020 5195.2  94.8 
- 1557.0 I      20.0 14.328 0.018 4516.5  74.3 
- 1621.0 I      20.0 14.381 0.021 4301.3  82.4 
- 1680.0 I      20.0 14.516 0.015 3798.4  52.1 
- 1744.0 I      20.0 14.572 0.028 3607.4  91.8 
- 1803.0 I      20.0 14.616 0.016 3464.2  50.7 
- 1862.0 I      20.0 14.686 0.021 3247.9  62.2 
- 1912.0 I      20.0 14.778 0.025 2984.0  67.9 
- 1972.0 I      20.0 14.807 0.025 2905.4  66.1 
- 2031.0 I      20.0 14.836 0.026 2828.8  66.9 
- 2158.0 I      20.0 14.958 0.021 2528.1  48.4 
- 2331.0 I      40.0 15.060 0.022 2301.4  46.2 
- 2575.5 I      40.0 15.174 0.023 2072.0  43.4 
- 2821.0 I      40.0 15.334 0.020 1788.1  32.6 
- 3065.5 I      40.0 15.439 0.022 1623.3  32.6 
- 3311.5 I      40.0 15.565 0.035 1445.4  45.9 
- 3555.5 I      40.0 15.700 0.031 1276.4  35.9 
- 3800.5 I      40.0 15.700 0.043 1276.4  49.6 
- 4046.5 I      40.0 15.788 0.031 1177.1  33.1 
- 4259.5 I      40.0 15.902 0.042 1059.7  40.2 
- 4379.5 I      40.0 15.892 0.043 1069.5  41.5 
- 4527.3 I      60.0 15.986 0.024  980.8  21.4 
- 5002.5 I      80.0 16.170 0.028  827.9  21.1 
- 5255.0 I      80.0 16.221 0.025  789.9  18.0 
- 5662.0 I      80.0 16.296 0.024  737.2  16.1 
- 6065.0 I      80.0 16.385 0.036  679.2  22.2 
- 6469.0 I      80.0 16.507 0.033  607.0  18.2 
- 6873.0 I      80.0 16.495 0.030  613.8  16.7 
- 8479.0 I     300.0 16.841 0.017  446.3   6.9 
- 9680.0 I     300.0 17.017 0.030  379.5  10.3 
-10924.0 I     300.0 17.204 0.019  319.4   5.5 
-12129.0 I     300.0 17.342 0.023  281.3   5.9 
-15195.0 I     300.0 17.648 0.022  212.2   4.3 
-16402.0 I     300.0 17.750 0.099  193.2  16.8
