--- grb/resources/grbs/GRB080319B/bloom_nickel.txt
@@ -1,48 +1,0 @@
-Table 5 
-Nickel Observations of GRB 080319B
-
-
-t_mid^a (s)	Filter	Exp. Time (s)	Mag.^b	Flux^b (muJy)	
-7154.7	B	360.0	17.365 +or- 0.005	467.7 +or- 2.1	
-8579.2	B	360.0	17.646 +or- 0.006	361.1 +or- 2.0	
-10013.4	B	360.0	17.866 +or- 0.008	294.8 +or- 2.2	
-11512.3	B	360.0	18.090 +or- 0.008	239.9 +or- 1.8	
-13015.9	B	360.0	18.248 +or- 0.011	207.4 +or- 2.1	
-14473.6	B	360.0	18.389 +or- 0.011	182.1 +or- 1.8	
-16436.8	B	360.0	18.587 +or- 0.015	151.8 +or- 2.1	
-17937.4	B	360.0	18.770 +or- 0.018	128.2 +or- 2.1	
-19398.1	B	360.0	18.858 +or- 0.015	118.2 +or- 1.6	
-20965.3	B	360.0	18.970 +or- 0.023	106.7 +or- 2.2	
-7444.0	V	300.0	17.251 +or- 0.005	464.1 +or- 2.1	
-8860.5	V	300.0	17.484 +or- 0.007	374.5 +or- 2.4	
-10291.9	V	300.0	17.727 +or- 0.007	299.4 +or- 1.9	
-11829.3	V	300.0	17.899 +or- 0.007	255.5 +or- 1.6	
-13294.2	V	300.0	18.062 +or- 0.007	219.9 +or- 1.4	
-14767.2	V	300.0	18.236 +or- 0.014	187.3 +or- 2.4	
-16714.4	V	300.0	18.424 +or- 0.012	157.5 +or- 1.7	
-18236.5	V	300.0	18.559 +or- 0.016	139.1 +or- 2.0	
-19700.4	V	300.0	18.637 +or- 0.013	129.5 +or- 1.5	
-22996.8	V	300.0	18.902 +or- 0.021	101.4 +or- 1.9	
-7782.0	R	300.0	16.991 +or- 0.006	496.1 +or- 2.7	
-9196.4	R	300.0	17.239 +or- 0.005	394.8 +or- 1.8	
-10635.4	R	300.0	17.435 +or- 0.008	329.6 +or- 2.4	
-12181.0	R	300.0	17.637 +or- 0.009	273.7 +or- 2.3	
-13652.0	R	300.0	17.799 +or- 0.007	235.7 +or- 1.5	
-15110.2	R	300.0	17.917 +or- 0.007	211.4 +or- 1.4	
-17071.9	R	300.0	18.107 +or- 0.011	177.5 +or- 1.8	
-18582.1	R	300.0	18.218 +or- 0.008	160.3 +or- 1.2	
-20035.8	R	300.0	18.308 +or- 0.017	147.5 +or- 2.3	
-23331.8	R	300.0	18.541 +or- 0.021	119.0 +or- 2.3	
-8122.0	I	300.0	16.797 +or- 0.004	464.7 +or- 1.7	
-9531.2	I	300.0	17.013 +or- 0.005	380.9 +or- 1.7	
-10995.4	I	300.0	17.189 +or- 0.005	323.9 +or- 1.5	
-12520.7	I	300.0	17.394 +or- 0.005	268.2 +or- 1.2	
-14003.3	I	300.0	17.537 +or- 0.009	235.1 +or- 1.9	
-15973.2	I	300.0	17.704 +or- 0.008	201.6 +or- 1.5	
-17412.9	I	300.0	17.817 +or- 0.011	181.6 +or- 1.8	
-18936.7	I	300.0	17.921 +or- 0.010	165.0 +or- 1.5	
-20447.9	I	300.0	18.045 +or- 0.014	147.2 +or- 1.9	
-23665.5	I	300.0	18.111 +or- 0.118	 138.5 +or- 14.3	
-Notes.
-a Exposure midtime, measured from the Swift trigger (UTC 06:12:49).
-b Observed value; not corrected for Galactic extinction.
