--- grb/resources/grbs/GRB080319B/bloom_uvot.txt
@@ -1,53 +1,0 @@
-Table 6 
-UVOT Observations of GRB 080319B
-
-
-t_mid^a (s)	Filter	Exp. Time (s)	Mag.^b	Flux^b (muJy)	
-6156.0	UVW1	196.6	16.146 +or- 0.043	343.2 +or- 13.3	
-17398.0	UVW1	885.6	17.645 +or- 0.042	86.3 +or- 3.3	
-28047.0	UVW1	427.9	18.316 +or- 0.086	46.5 +or- 3.5	
-45393.0	UVW1	427.9	18.909 +or- 0.121	26.9 +or- 2.8	
-75628.0	UVW1	200.6	19.266 +or- 0.213	19.4 +or- 3.5	
-5542.0	UVW2	196.6	16.192 +or- 0.046	239.6 +or- 9.9	
-10709.0	UVW2	885.6	17.231 +or- 0.036	92.0 +or- 3.0	
-22273.0	UVW2	266.3	18.394 +or- 0.116	31.5 +or- 3.2	
-28931.0	UVW2	856.1	18.631 +or- 0.073	25.3 +or- 1.6	
-46279.0	UVW2	856.1	19.228 +or- 0.102	14.6 +or- 1.3	
-76050.0	UVW2	401.3	19.888 +or- 0.220	8.0 +or- 1.5	
-5951.0	UVM2	196.6	15.923 +or- 0.052	284.4 +or- 13.3	
-12528.0	UVM2	334.6	17.144 +or- 0.071	92.4 +or- 5.8	
-16491.0	UVM2	885.6	17.493 +or- 0.052	67.0 +or- 3.1	
-30028.0	UVM2	611.2	18.381 +or- 0.101	29.6 +or- 2.6	
-47376.0	UVM2	610.0	19.039 +or- 0.144	16.1 +or- 2.0	
-631.0	U	19.4	12.474 +or- 0.028^x	19697.0 +or- 501.5^x	
-785.0	U	19.4	12.968 +or- 0.030	12496.8 +or- 340.6	
-1430.0	U	19.4	14.108 +or- 0.042	4373.2 +or- 165.9	
-1591.0	U	19.5	14.342 +or- 0.046	3525.3 +or- 146.2	
-4926.0	U	196.6	16.015 +or- 0.029	755.1 +or- 19.9	
-6362.0	U	196.6	16.329 +or- 0.033	565.5 +or- 16.9	
-18304.0	U	295.0	17.864 +or- 0.056	137.5 +or- 6.9	
-18609.0	U	231.2	17.886 +or- 0.064	134.8 +or- 7.7	
-28486.0	U	213.4	18.498 +or- 0.089	76.7 +or- 6.0	
-45833.0	U	213.4	19.435 +or- 0.148	32.4 +or- 4.1	
-656.0	B	9.6	13.334 +or- 0.042^x	19160.2 +or- 727.0^x	
-811.0	B	9.6	13.913 +or- 0.046	11240.9 +or- 466.3	
-1454.0	B	19.4	15.012 +or- 0.046	4085.1 +or- 169.5	
-1616.0	B	19.5	15.291 +or- 0.052	3159.4 +or- 147.7	
-5131.0	B	196.6	16.962 +or- 0.033	678.0 +or- 20.3	
-6567.0	B	196.6	17.350 +or- 0.039	474.2 +or- 16.7	
-28708.0	B	213.4	19.417 +or- 0.112	70.7 +or- 6.9	
-46055.0	B	213.3	20.036 +or- 0.171	40.0 +or- 5.8	
-75943.0	B	100.2	20.441 +or- 0.316	27.5 +or- 6.9	
-711.0	V	19.5	13.373 +or- 0.040^x	16512.0 +or- 597.2^x	
-975.0	V	393.5	14.337 +or- 0.013^x	6795.2 +or- 80.9^x	
-1518.0	V	19.4	15.025 +or- 0.076	3605.8 +or- 243.8	
-1680.0	V	19.5	15.127 +or- 0.081	3282.5 +or- 236.0	
-5746.0	V	196.6	16.955 +or- 0.058	609.5 +or- 31.7	
-11616.0	V	295.0	17.948 +or- 0.078	244.2 +or- 16.9	
-11919.0	V	295.1	17.895 +or- 0.078	256.4 +or- 17.8	
-12222.0	V	295.0	18.041 +or- 0.082	224.2 +or- 16.3	
-29806.0	V	213.4	19.491 +or- 0.256	59.0 +or- 12.4	
-Notes.
-a Exposure midtime, measured from the Swift trigger (UTC 06:12:49).
-b Observed value; not corrected for Galactic extinction.
-x Point not used in modeling.
