--- grb/resources/grbs/GRB080319B/doi10.1051-0004-6361-200811135.txt
@@ -1,96 +1,0 @@
-2008 Mar 19.4616		0.2027	1200		none 	>15.5			0.22mSR-22
-2008 Mar 19.7642		0.5052	600			R		19.07	0.19	0.7mAZT-8
-2008 Mar 19.7721		0.5132	600			R		19.26	0.27	0.7mAZT-8
-2008 Mar 19.7800		0.5211	600			R		19.74	0.50	0.7mAZT-8
-2008 Mar 19.7911		0.5322	600			R		19.96	0.39	0.7mAZT-8
-2008 Mar 19.7974		0.5385	600			R		19.60	0.34	0.7mAZT-8
-2008 Mar 19.8055		0.5466	900			R		19.60	0.20	0.7mAZT-8
-2008 Mar 19.8174		0.5585	900			R		19.40	0.15	0.7mAZT-8
-2008 Mar 19.8285		0.5696	900			R		19.89	0.28	0.7mAZT-8
-2008 Mar 19.8470		0.5881	900			R		19.82	0.23	0.7mAZT-8
-2008 Mar 19.8557		0.5968	900			R		20.06	0.19	0.7mAZT-8
-2008 Mar 19.8630		0.6041	900			R		19.88	0.23	0.7mAZT-8
-2008 Mar 19.8888		0.6299	720			R		20.00	0.27	0.7mAZT-8
-2008 Mar 19.8980		0.6391	720			R		19.52	0.16	0.7mAZT-8
-2008 Mar 19.9071		0.6482	720			R		20.09	0.28	0.7mAZT-8
-2008 Mar 19.9139		0.6550	720			R		19.80	0.24	0.7mAZT-8
-2008 Mar 19.9234		0.6645	900			R		19.88	0.27	0.7mAZT-8
-2008 Mar 19.9354		0.6765	900			R		20.19	0.40	0.7mAZT-8
-2008 Mar 19.9497		0.6908	900			R		19.54	0.15	0.7mAZT-8
-2008 Mar 19.9625		0.7036	900			R		20.16	0.44	0.7mAZT-8
-2008 Mar 19.9752		0.7163	720			R		20.47	0.25	0.7mAZT-8
-2008 Mar 19.9843		0.7254	960			R		20.00	0.37	0.7mAZT-8
-2008 Mar 19.9959		0.7370	900			R		20.03	0.37	0.7mAZT-8
-2008 Mar 20.0127		0.7538	720			R		20.01	0.22	0.7mAZT-8
-2008 Mar 20.0196		0.7607	720			R		19.83	0.21	0.7mAZT-8
-2008 Mar 20.0664		0.8075	720			R		19.94	0.23	0.7mAZT-8
-2008 Mar 20.0755		0.8166	720			R		20.22	0.26	0.7mAZT-8
-2008 Mar 20.0848		0.8259	900			R		20.28	0.24	0.7mAZT-8
-2008 Mar 20.0959		0.8370	900			R		19.74	0.16	0.7mAZT-8
-2008 Mar 20.1069		0.8480	600			R		19.56	0.20	0.7mAZT-8
-2008 Mar 19.8319		0.5730	600			V		19.60	0.70	0.7mAZT-8
-2008 Mar 20.0870		0.8351	1200		I		19.57	0.08	0.8mIAC80
-2008 Mar 19.9924		0.7457	4×300		R		19.82	0.07	0.8mIAC80
-2008 Mar 20.0557		0.8003	600			R		19.94	0.11	0.8mIAC80
-2008 Mar 20.1067		0.8548	2×600		R		20.15	0.08	0.8mIAC80
-2008 Mar 20.1294		0.8778	2×600		R		20.14	0.09	0.8mIAC80
-2008 Mar 20.2454		0.9972	3×600		R		20.50	0.08	0.8mIAC80
-2008 Mar 22.0730		2.8355	6×600		R		22.03	0.27	0.8mIAC80
-2008 Mar 20.0252		0.7814	(1200+600)	V		20.39	0.10	0.8mIAC80
-2008 Mar 20.0648		0.8165	3×600		B		20.89	0.12	0.8mIAC80
-2008 Mar 20.1703		0.9149	600			B		20.75	0.20	0.8mIAC80
-2008 Mar 19.9361		0.6985	120			r		19.56	0.17	1.2mMercator
-2008 Mar 19.9613		0.7031	120			r		19.28	0.19	1.2mMercator
-2008 Mar 19.9631		0.7049	120			r		19.74	0.15	1.2mMercator
-2008 Mar 19.9649		0.7067	120			r		19.37	0.11	1.2mMercator
-2008 Mar 19.9667		0.7085	120			r		19.93	0.15	1.2mMercator
-2008 Mar 19.9681		0.7095	60			r		19.53	0.11	1.2mMercator
-2008 Mar 19.9691		0.7106	60			r		20.07	0.21	1.2mMercator
-2008 Mar 19.9702		0.7117	60			r		19.41	0.10	1.2mMercator
-2008 Mar 19.9713		0.7127	60			r		19.66	0.20	1.2mMercator
-2008 Mar 19.9723		0.7138	60			r		19.82	0.15	1.2mMercator
-2008 Mar 19.9734		0.7148	60			r		19.68	0.09	1.2mMercator
-2008 Mar 19.9745		0.7159	60			r		19.61	0.09	1.2mMercator
-2008 Mar 19.9759		0.7177	120			r		19.57	0.12	1.2mMercator
-2008 Mar 19.9777		0.7195	120			r		19.54	0.09	1.2mMercator
-2008 Mar 19.9794		0.7212	120			r		19.62	0.07	1.2mMercator
-2008 Mar 19.9812		0.7230	120			r		19.83	0.09	1.2mMercator
-2008 Mar 19.9830		0.7248	120			r		19.58	0.06	1.2mMercator
-2008 Mar 19.9848		0.7266	120			r		19.65	0.07	1.2mMercator
-2008 Mar 19.9865		0.7283	120			r		19.67	0.07	1.2mMercator
-2008 Mar 19.9883		0.7301	120			r		19.78	0.08	1.2mMercator
-2008 Mar 19.9900		0.7318	120			r		19.51	0.07	1.2mMercator
-2008 Mar 19.9918		0.7336	120			r		19.66	0.08	1.2mMercator
-2008 Mar 19.9936		0.7354	120			r		19.47	0.05	1.2mMercator
-2008 Mar 19.9953		0.7371	120			r		19.62	0.06	1.2mMercator
-2008 Mar 19.9971		0.7389	120			r		19.46	0.05	1.2mMercator
-2008 Mar 19.9989		0.7407	120			r		19.62	0.06	1.2mMercator
-2008 Mar 20.1446		0.8871	120			r		19.76	0.07	1.2mMercator
-2008 Mar 20.1464		0.8888	120			r		19.95	0.08	1.2mMercator
-2008 Mar 20.1481		0.8906	120			r		20.14	0.10	1.2mMercator
-2008 Mar 20.1500		0.8925	120			r		19.78	0.07	1.2mMercator
-2008 Mar 20.1517		0.8942	120			r		20.04	0.11	1.2mMercator
-2008 Mar 20.1535		0.8960	120			r		19.98	0.10	1.2mMercator
-2008 Mar 20.1554		0.8979	120			r		19.85	0.08	1.2mMercator
-2008 Mar 20.1572		0.8997	120			r		19.85	0.12	1.2mMercator
-2008 Mar 20.1589		0.9014	120			r		19.87	0.13	1.2mMercator
-2008 Mar 20.1607		0.9032	120			r		19.87	0.11	1.2mMercator
-2008 Mar 20.1625		0.9049	120			r		19.92	0.09	1.2mMercator
-2008 Mar 20.1642		0.9067	120			r		19.96	0.11	1.2mMercator
-2008 Mar 20.1660		0.9085	120			r		20.04	0.10	1.2mMercator
-2008 Mar 20.1678		0.9103	120			r		19.84	0.08	1.2mMercator
-2008 Mar 20.1695		0.9120	120			r		19.73	0.09	1.2mMercator
-2008 Mar 20.1713		0.9138	120			r		19.84	0.11	1.2mMercator
-2008 Mar 20.1731		0.9156	120			r		19.92	0.10	1.2mMercator
-2008 Mar 20.1748		0.9173	120			r		20.01	0.12	1.2mMercator
-2008 Mar 20.1766		0.9191	120			r		19.91	0.10	1.2mMercator
-2008 Mar 20.1784		0.9208	120			r		20.08	0.09	1.2mMercator
-2008 Mar 20.1801		0.9226	120			r		19.71	0.07	1.2mMercator
-2008 Mar 20.1819		0.9244	120			r		19.87	0.09	1.2mMercator
-2008 Mar 20.1837		0.9261	120			r		19.83	0.07	1.2mMercator
-2008 Mar 20.1854		0.9279	120			r		19.84	0.09	1.2mMercator
-2008 Mar 21.8618		2.6029	5760		I		21.22	0.15	1.5mAZT-22
-2008 Mar 21.8781		2.6192	5610		R		21.81	0.26	1.5mAZT-22
-2008 Apr 07.0086		18.7479	7200		R		>23.6			1.5mAZT-22
-2008 Mar 19.9179		0.6590	3×600		R		19.66	0.09 	2.0mIGO
-2008 Mar 27.9511		8.6524	28×120		R		>23.0			2.5mNOT
