--- grb/resources/grbs/GRB090424/ping_optical.txt
@@ -1,159 +1,0 @@
-Title: GRB081007 and GRB090424: the surrounding medium, outflows and supernovae 
-Authors: Jin Z.-P., Covino S., Della Valle M., Ferrero P., Fugazza D.,
-         Malesani D., Melandri A., Pian E., Salvaterra R., Bersier D., 
-         Campana S., Cano Z., D'Avanzo P., Fynbo J.P.U., Guidorzi C.,
-         Haislip J.B., Hjorth J., LaCluyze A.P., Marconi G., Mazzali P.A., 
-         Piranomonte S., Reichart D.E., Tagliaferri G., Tanvir N.R., 
-         Valenti S., Vergani S.D., Vestrand T., Walker E.S., Wozniak P.
-Table: Optical observations of GRB 090424
-================================================================================
-Byte-by-byte Description of file: apj480429t3_mrt.txt
---------------------------------------------------------------------------------
-   Bytes Format Units   Label     Explanations
---------------------------------------------------------------------------------
-   1-  4 A4     ---     Inst      Instrument used in the observation
-   6- 10 A5     ---     Filter    Filter used in the observation
-  12- 20 F9.2   min     Mid-time  Time of mid-exposure (1)
-  22- 26 F5.2   min     Exp       Exposure time
-  28- 32 F5.2   mag     mag       Observed AB magnitude in Filter (2)
-  34- 38 F5.3   mag   e_mag       Error in mag
---------------------------------------------------------------------------------
-Note (1): Since trigger time: 2009 April 24 at 14:12:09 UT.
-Note (2): Galactic extinction has not been removed.
---------------------------------------------------------------------------------
-UVOT UVW2      13.22  0.32 17.54 0.28  
-UVOT UVW2      19.94  0.32 17.66 0.29  
-UVOT UVW2     133.50  2.02 19.77 0.38  
-UVOT UVM2      11.17  0.32 17.39 0.34  
-UVOT UVM2     320.28 13.80 20.04 0.22  
-UVOT UVM2     603.99 14.76 20.59 0.31  
-UVOT UVW1      11.58  0.33 16.77 0.19  
-UVOT UVW1      18.30  0.32 17.93 0.34  
-UVOT UVW1      21.17  0.32 17.49 0.28  
-UVOT UVW1      24.04  0.32 17.69 0.31  
-UVOT UVW1      26.94  0.33 18.49 0.48  
-UVOT UVW1      29.82  0.32 18.20 0.41  
-UVOT UVW1      32.71  0.33 18.66 0.52  
-UVOT UVW1      35.58  0.33 17.56 0.28  
-UVOT UVW1      94.17  3.28 18.79 0.19  
-UVOT UVW1     417.97 11.20 20.52 0.33  
-UVOT u          5.15  0.17 16.00 0.13  
-UVOT u          5.32  0.17 16.29 0.15  
-UVOT u          5.48  0.17 16.25 0.15  
-UVOT u          5.65  0.17 16.23 0.14  
-UVOT u          5.82  0.17 16.36 0.15  
-UVOT u          5.98  0.17 16.35 0.15  
-UVOT u          6.15  0.17 16.73 0.19  
-UVOT u          6.32  0.17 16.46 0.16  
-UVOT u          6.48  0.17 16.47 0.16  
-UVOT u          6.65  0.17 16.75 0.19  
-UVOT u          6.82  0.17 16.38 0.15  
-UVOT u          6.98  0.17 16.87 0.20  
-UVOT u          7.15  0.17 16.76 0.19  
-UVOT u          7.32  0.17 16.63 0.18  
-UVOT u          7.48  0.17 17.05 0.22  
-UVOT u          7.65  0.17 16.73 0.18  
-UVOT u          7.82  0.17 16.68 0.18  
-UVOT u          7.98  0.17 16.55 0.17  
-UVOT u          8.15  0.17 17.00 0.22  
-UVOT u          8.32  0.17 16.91 0.20  
-UVOT u          8.48  0.17 16.83 0.19  
-UVOT u          8.65  0.17 16.63 0.17  
-UVOT u          8.82  0.17 17.29 0.25  
-UVOT u          8.98  0.17 16.78 0.19  
-UVOT u          9.15  0.16 16.84 0.20  
-UVOT u         11.82  0.09 17.07 0.30  
-UVOT u         11.98  0.17 17.19 0.23  
-UVOT u         12.15  0.08 17.30 0.37  
-UVOT u         18.70  0.33 17.41 0.19  
-UVOT u         24.45  0.33 17.53 0.20  
-UVOT u         27.35  0.33 17.73 0.23  
-UVOT u         33.12  0.32 18.65 0.42  
-UVOT u         35.99  0.32 18.04 0.28  
-UVOT u         96.14  0.44 18.01 0.24  
-UVOT u        103.97  3.28 19.12 0.19  
-UVOT u        500.30  4.92 20.01 0.31  
-UVOT u        505.37  4.92 20.23 0.37  
-UVOT u        510.43  4.92 19.86 0.27  
-UVOT b          9.51  0.33 17.32 0.14  
-UVOT b         12.39  0.33 17.46 0.14  
-UVOT b         19.12  0.32 17.84 0.18  
-UVOT b         21.98  0.32 18.06 0.21  
-UVOT b         24.89  0.33 18.30 0.24  
-UVOT b         27.77  0.33 18.56 0.29  
-UVOT b         30.64  0.33 18.36 0.26  
-UVOT b         33.53  0.32 18.75 0.34  
-UVOT b         36.40  0.32 18.07 0.21  
-UVOT b        107.40  3.28 19.79 0.24  
-UVOT b        229.69  2.41 20.13 0.37  
-UVOT b        515.53  4.92 20.29 0.30  
-UVOT b        519.06  1.92 20.80 0.74  
-UVOT v          1.33  0.16 13.82 0.07  
-UVOT v         10.74  0.33 16.88 0.19  
-UVOT v         13.64  0.33 17.19 0.24  
-UVOT v         17.48  0.33 17.65 0.31  
-UVOT v         20.35  0.33 17.97 0.39  
-UVOT v         23.21  0.33 17.89 0.37  
-UVOT v         26.11  0.33 17.79 0.35  
-UVOT v         29.00  0.33 18.75 0.72  
-UVOT v         31.89  0.32 17.63 0.32  
-UVOT v         34.76  0.32 18.35 0.53  
-UVOT v         37.66  0.32 17.58 0.30  
-UVOT v        200.71  4.92 19.55 0.35  
-UVOT v        203.53  0.52 19.77 1.33  
-UVOT White      1.62  0.17 15.03 0.05  
-UVOT White      1.78  0.17 15.39 0.05  
-UVOT White      1.95  0.17 15.27 0.05  
-UVOT White      2.12  0.17 15.46 0.05  
-UVOT White      2.28  0.17 15.57 0.06  
-UVOT White      2.45  0.17 15.43 0.05  
-UVOT White      2.62  0.17 15.55 0.06  
-UVOT White      2.78  0.17 15.66 0.06  
-UVOT White      2.95  0.17 15.70 0.06  
-UVOT White      3.12  0.17 15.77 0.06  
-UVOT White      3.28  0.17 15.87 0.06  
-UVOT White      3.45  0.17 15.99 0.06  
-UVOT White      3.62  0.17 16.11 0.06  
-UVOT White      3.78  0.17 16.06 0.06  
-UVOT White      3.95  0.16 16.23 0.07  
-UVOT White      9.78  0.12 17.26 0.12  
-UVOT White      9.95  0.17 17.55 0.13  
-UVOT White     10.12  0.04 17.77 0.29  
-UVOT White     12.62  0.07 17.51 0.19  
-UVOT White     12.78  0.17 17.66 0.13  
-UVOT White     12.95  0.10 18.00 0.21  
-UVOT White     19.45  0.17 18.37 0.21  
-UVOT White     19.62  0.15 18.11 0.18  
-UVOT White     22.28  0.14 18.28 0.21  
-UVOT White     22.45  0.17 18.07 0.17  
-UVOT White     25.28  0.33 18.31 0.14  
-UVOT White     28.17  0.32 18.06 0.13  
-UVOT White     31.06  0.33 18.14 0.13  
-UVOT White     33.93  0.32 18.29 0.15  
-UVOT White     36.81  0.33 18.16 0.13  
-UVOT White    110.47  2.61 19.26 0.11  
-LT   r'       453.22 30.00 19.67 0.02  
-LT   r'       573.28 30.00 19.91 0.08  
-LT   i'       499.33 30.00 19.66 0.04  
-FTN  r'     25435.85 30.00 21.86 0.11  
-FTN  r'    480809.88 30.00 21.65 0.11  
-FTN  r'    482086.00 30.00 22.03 0.15  
-FTN  r'    483666.35 30.00 21.78 0.15  
-VLT  r'      9360.01  6.00 22.00 0.10  
-VLT  r'     26619.71  6.00 22.08 0.15  
-VLT  r'     38151.23  6.00 22.13 0.08  
-VLT  r'     71316.72  6.00 21.90 0.35  
-VLT  r'    104267.52  3.00 22.50 0.21  
-VLT  i'      9344.68  4.00 21.65 0.08  
-VLT  i'      9349.57  4.00 21.56 0.08  
-VLT  i'      9354.13  4.00 21.66 0.08  
-VLT  i'     26627.66  4.00 21.80 0.10  
-VLT  i'     26634.29  4.00 21.79 0.09  
-VLT  i'     38156.93  4.00 21.78 0.06  
-VLT  i'     38161.57  4.00 21.77 0.07  
-VLT  i'     71322.34  4.00 21.88 0.14  
-VLT  i'     71326.81  4.00 21.85 0.16  
-VLT  i'    104238.34  4.00 21.86 0.13  
-VLT  i'    104247.08  4.00 21.85 0.13  
-CAHA J     1589029.99 212.00 21.60 0.11
-CAHA H     1589080.06 47.00  21.21 0.32
