--- grb/results/example.txt
@@ -1,18 +1,0 @@
-###################################################################################################################
-# GRB		TYPE	GAMMA_B				VIEWING				OPENING				SIG		SOURCE
-###################################################################################################################
-GRB120224A	LONG	1.810 +0.56 -0.45	0.55 +0.11 -0.13	0.55 +0.17 -0.14	4.23	MNRAS 530, 2877-2884 (2024)	
-GRB120401A	LONG	10.40 +1.14 -1.75	0.03 +0.02 -0.01	0.10 +0.01 -0.02	3.00	MNRAS 530, 2877-2884 (2024)	
-GRB120711A	LONG	2.890 +0.44 -0.46	0.19 +0.04 -0.05	0.35 +0.05 -0.06	3.80	MNRAS 530, 2877-2884 (2024)	
-GRB120811C	LONG	1.990 +0.89 -0.49	0.34 +0.10 -0.10	0.50 +0.22 -0.12	3.40	MNRAS 530, 2877-2884 (2024)	
-GRB120907A	LONG	1.350 +0.11 -0.10	0.49 +0.04 -0.07	0.74 +0.06 -0.05	7.00	MNRAS 530, 2877-2884 (2024)	
-GRB120909A	LONG	2.480 +0.67 -0.52	0.24 +0.08 -0.08	0.40 +0.11 -0.08	3.00	MNRAS 530, 2877-2884 (2024)	
-GRB120922A	LONG	1.440 +0.31 -0.17	0.42 +0.07 -0.12	0.69 +0.15 -0.08	3.50	MNRAS 530, 2877-2884 (2024)	
-GRB121011A	LONG	6.480 +0.74 -1.15	0.18 +0.01 -0.01	0.15 +0.02 -0.03	18.0	MNRAS 530, 2877-2884 (2024)	
-GRB121027A	LONG	5.030 +0.07 -2.80	0.43 +0.19 -0.04	0.20 +0.00 -0.11	10.75	MNRAS 530, 2877-2884 (2024)	
-GRB121201A	LONG	1.890 +0.57 -0.35	0.30 +0.05 -0.10	0.53 +0.16 -0.10	3.00	MNRAS 530, 2877-2884 (2024)	
-GRB121229A	LONG	6.840 +0.95 -0.79	0.13 +0.01 -0.01	0.15 +0.02 -0.02	13.0	MNRAS 530, 2877-2884 (2024)	
-###################################################################################################################
-# Filtered by viewing angle confidence greater than 3 inclusive
-# Filtered by date between Jan 1, 2012 and Jan 1, 2013 inclusive
-###################################################################################################################
