+++ requirements.txt
@@ -13,3 +13,6 @@ PyQt6-Qt6==6.7.2
 PyQt6_sip==13.8.0
 python-dateutil==2.9.0.post0
 six==1.16.0
+
+scipy~=1.14.1
+pandas~=2.2.2
